fixed jest walletinfo issue
